patitas/requirements.txt
django-crispy-forms
social-auth-app-django
django-admin-interface
django-allauth